Move point precompile tests to barycentric file
